feat: add wasp API endpoint
