+++ packages/shared/src/lib/core/network/classes/iscp-chain.class.ts
@@ -9,20 +9,21 @@ import { ChainConfiguration, ChainMetadata, Web3Provider } from '../types'
 export class IscpChain implements IChain {
     private readonly _provider: Web3Provider
     private readonly _configuration: IIscpChainConfiguration
+    private readonly _chainApi: string
 
     private _metadata: IIscpChainMetadata | undefined
-
     constructor(payload: IIscpChainConfiguration) {
         try {
             /**
              * NOTE: We can assume that the data inside this payload has already
              * been validated.
              */
-            const { aliasAddress, iscpEndpoint } = payload
+            const { aliasAddress, rpcEndpoint, apiEndpoint } = payload
             const evmJsonRpcPath = this.buildEvmJsonRpcPath(aliasAddress)
 
-            this._provider = new Web3(`${iscpEndpoint}/${evmJsonRpcPath}`)
+            this._provider = new Web3(`${rpcEndpoint}/${evmJsonRpcPath}`)
             this._configuration = payload
+            this._chainApi = `${apiEndpoint}v1/chains/${aliasAddress}`
         } catch (err) {
             console.error(err)
             throw new Error('Failed to construct ISCP Chain!')
@@ -70,14 +71,34 @@ export class IscpChain implements IChain {
      * node URL). See here for more: https://github.com/iotaledger/wasp/issues/2385
      */
     private async fetchChainMetadata(): Promise<IIscpChainMetadata> {
-        const { aliasAddress, iscpEndpoint } = this._configuration
-        const chainMetadataUrl = `${iscpEndpoint}/v1/chains/${aliasAddress}`
-        const response = await fetch(chainMetadataUrl)
+        const response = await fetch(this._chainApi)
         return (await response.json()) as IIscpChainMetadata
     }
 
     async getLatestBlock(): Promise<IBlock> {
         const number = await this._provider.eth.getBlockNumber()
         return this._provider.eth.getBlock(number)
+    }
+
+    async getGasEstimate(hex: string): Promise<unknown> {
+        const URL = `${this._chainApi}/estimategas-onledger`
+        const body = JSON.stringify({ outputBytes: hex })
+
+        const requestInit: RequestInit = {
+            method: 'POST',
+            body,
+            headers: {
+                Accept: 'application/json',
+                'Content-Type': 'application/json',
+            },
+        }
+
+        const response = await fetch(URL, requestInit)
+        if (response.status === 200) {
+            const data = await response.json()
+            const gasBurned = Number(data.gasBurned as string)
+            const gasFeeCharged = Number(data.gasFeeCharged as string)
+            return { gasBurned, gasFeeCharged }
+        }
     }
 }
