+++ packages/shared/src/lib/core/network/constants/default-chain-configurations.constant.ts
@@ -12,6 +12,7 @@ export const DEFAULT_CHAIN_CONFIGURATIONS: Readonly<{ [id in NetworkId]?: ChainC
         coinType: DEFAULT_COIN_TYPE[SupportedNetworkId.ShimmerEvm],
         aliasAddress: 'smr1prxvwqvwf7nru5q5xvh5thwg54zsm2y4wfnk6yk56hj3exxkg92mx20wl3s',
         rpcEndpoint: 'https://json-rpc.evm.shimmer.network/',
+        apiEndpoint: 'https://api.evm.shimmer.network/',
         explorerUrl: 'https://explorer.evm.shimmer.network/',
     },
     [SupportedNetworkId.Testnet]: {
@@ -23,6 +24,7 @@ export const DEFAULT_CHAIN_CONFIGURATIONS: Readonly<{ [id in NetworkId]?: ChainC
         coinType: DEFAULT_COIN_TYPE[SupportedNetworkId.TestnetEvm],
         aliasAddress: 'rms1ppp00k5mmd2m8my8ukkp58nd3rskw6rx8l09aj35984k74uuc5u2cywn3ex',
         rpcEndpoint: 'https://json-rpc.evm.testnet.shimmer.network/',
+        apiEndpoint: 'https://api.evm.testnet.shimmer.network/',
         explorerUrl: 'https://explorer.evm.testnet.shimmer.network/',
     },
 }
